Adjusted the css file to get desired outcome
Set the .header to display: flex; and adjusted the left-links to margin-right: auto; to get the left links to align to the far left of the logo, and I adjusted the right-links to margin-left: auto; in order to push the right links to the far right. I struggled witht with the padding of the LOGO, so that it would reveal the header background but was unsuccessful. I tried placing the gaps for the links within the parent containers for each <ul> but was unsuccessful. After multiple failed attempts I looked at the solution to find that I needed to place the display: flex; inside both the header container and the <ul> element container. To get the desired padding for the logo I needed to adjust the padding for the header container which would expose the background color around the logo since it is part of the content box being push away from the border with padding. Eliminated the space before a list by adding margin: 0; and padding: 0; and adding space between the items with gap: 8px;