Applied changes to file to create modern header
Applied display: flex; to the header, ul, and .left, .right containers.(header parent container, and ul, logo and left,right child flex items). Justify-content: space between; under .header in css to create even space between items. Set .header padding to 8px, and for the ul, margin and paddign are set to 0 to eliminate the default space and indention, then added gap spacing of 16px. The .left and .right containers were adjust with align-items: center; to center them along the cross axis within the .header. with a gap: 16px; to match the gap spacing (16px) of the <ul> items for a clean look balanced look.