Completed flex Layout
Had to rely on the solution code after not being able to get things in the right place, I would say about 40% to 50% of my code was correct, some code was not even needing. The sidebar column gave me the most trouble, and I got tunnel vision on the container next to the sidebar and forgot to use padding to the conatainer and padding on the indvidual .card items. I was right to use flex-wrap: wrap;. I was using flex-direction: column more than I needed to and couldnt get the sidebar column to stretch to the bottom of the page, after looking at solution, I only needed to use column flex box on the body. Learned alot from these exercises.